Manually redo changes that git apparently discarded
