+++ Core.ApplicationServices/IUserService.cs
@@ -33,5 +33,6 @@ namespace Core.ApplicationServices
         bool IsEmailInUse(string email);
         Result<User, OperationError> GetUserByEmail(Guid organizationUuid, string email);
         Result<User, OperationError> GetUserByUuid(Guid userUuid);
-    }
+        Maybe<OperationError> DeleteUserByOrganizationUuid(Guid userUuid, Guid scopedToOrganizationUuid);
+}
 }
