+++ Tests.Unit.Core.ApplicationServices/Tests.Unit.Core.csproj
@@ -227,7 +227,6 @@
     <Compile Include="ApplicationServices\SystemUsage\ItSystemUsageWriteServiceTest.cs" />
     <Compile Include="ApplicationServices\UIModuleCustomizationTest\UIModuleCustomizationTests.cs" />
     <Compile Include="ApplicationServices\UserRightsServiceTest.cs" />
-    <Compile Include="ApplicationServices\UserServiceTest.cs" />
     <Compile Include="ApplicationServices\Users\UserServiceTest.cs" />
     <Compile Include="ApplicationServices\Users\UserWriteServiceTest.cs" />
     <Compile Include="BackgroundJobs\ScheduleFkOrgUpdatesBackgroundJobTest.cs" />
